refactor: GroupScheduleToVoteScheduleTransform -> GroupScheduleTypeTransform 안으로 코드 이동
